+++ changelog.md
@@ -1,5 +1,9 @@
 # webpack-watch-server change log
 
+## next
+
+- Added a new `--config` CLI argument for a custom Webpack config file path, via [#1](https://github.com/jaydenseric/webpack-watch-server/pull/1).
+
 ## 1.1.0
 
 - Removed the redundant `fined` dependency.